+++ changelog.md
@@ -1,5 +1,6 @@
 
 ## TODO
+	try enhance MIDI matching in svg notation
 	staffNotation
 		remove note time from staff document
 	lilyInterpreter